frontend: Complete guest limited functionalities

+++ frontend/src/components/EditModal/EditModal.jsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react'
 import './EditModal.css'
 import { useData } from '../../context/DataContext';
 import axios from 'axios';
+import { useAuth } from '../../context/AuthContext';
 
 
 const EditModal = ({ selectedRefrigerator, setOpenModal }) => {
@@ -9,7 +10,8 @@ const EditModal = ({ selectedRefrigerator, setOpenModal }) => {
     const [refriName, setRefriName] = useState(selectedRefrigerator.refri_name)
     const [selectedAmount, setSelectedAmount] = useState(selectedRefrigerator.total_capacity)
     const [hideWarning, setHideWarning] = useState(true)
-    const { setRefrigerators, setLoadingRefrigerators, refriAmount, setRefriAmount } = useData();
+    const { setRefrigerators, setLoadingRefrigerators, updateMockRefrigerator, refriAmount, setRefriAmount } = useData();
+    const { user } = useAuth() 
 
     const handleAmountChange = (event) => {
         setSelectedAmount(+event.target.value)
@@ -28,41 +30,46 @@ const EditModal = ({ selectedRefrigerator, setOpenModal }) => {
 
     const saveChanges = async () => {
         try {
+            const url = process.env.REACT_APP_BACKEND_URL
             const updatedRefrigerator = {
                 refri_name: refriName,
                 total_capacity: selectedAmount,
                 status: selectedRefrigerator.status
             }
-            const url = process.env.REACT_APP_BACKEND_URL
-            await axios.put(`${url}/refrigerators/${selectedRefrigerator._id}`, updatedRefrigerator)
-            const amount = selectedAmount - selectedRefrigerator.total_capacity
-            if(amount > 0){
-                for(let i = 0; i < amount; i++){
-                    let startPosition = Math.floor(selectedRefrigerator.total_capacity / 2) + 1
-                    const newBucket = {
-                        refrigerator_id: selectedRefrigerator._id,
-                        flavor_id: null,
-                        position: startPosition + Math.floor(i / 2), 
-                        side: i % 2
-                    }
-                    await axios.post(`${url}/buckets`, newBucket)
-                }
-            }
-            if(amount < 0){
-                const selectedRefrigeratorCopy = selectedRefrigerator
-                const sortedBuckets = selectedRefrigeratorCopy.buckets.sort((a, b) => b.position - a.position);
-                for (let i = 0; i < (amount * -1); i++) {
-                    if (i < sortedBuckets.length) {
-                        await axios.delete(`${url}/buckets/${sortedBuckets[i]._id}`)
-                    } else {
-                        break;
+            if(user){
+                await axios.put(`${url}/refrigerators/${selectedRefrigerator._id}`, updatedRefrigerator)
+                const amount = selectedAmount - selectedRefrigerator.total_capacity
+                if(amount > 0){
+                    for(let i = 0; i < amount; i++){
+                        let startPosition = Math.floor(selectedRefrigerator.total_capacity / 2) + 1
+                        const newBucket = {
+                            refrigerator_id: selectedRefrigerator._id,
+                            flavor_id: null,
+                            position: startPosition + Math.floor(i / 2), 
+                            side: i % 2
+                        }
+                        await axios.post(`${url}/buckets`, newBucket)
                     }
                 }
+                if(amount < 0){
+                    const selectedRefrigeratorCopy = selectedRefrigerator
+                    const sortedBuckets = selectedRefrigeratorCopy.buckets.sort((a, b) => b.position - a.position);
+                    for (let i = 0; i < (amount * -1); i++) {
+                            if (i < sortedBuckets.length) {
+                                    await axios.delete(`${url}/buckets/${sortedBuckets[i]._id}`)
+                            } else {
+                                break;
+                            }
+                    }
+                }
+                setOpenModal(false)
+                const refrigeratorsFetch = await axios.get(`${url}/refrigerators`);
+                setRefrigerators(refrigeratorsFetch.data)
+            
+            }else{
+                updateMockRefrigerator(updatedRefrigerator.refri_name, selectedRefrigerator._id)
+                setOpenModal(false)
             }
-            setOpenModal(false)
-            const refrigeratorsFetch = await axios.get(`${url}/refrigerators`);
-            setRefrigerators(refrigeratorsFetch.data)
-
         } catch (error) {
             console.log(error);
         }
@@ -87,7 +94,7 @@ const EditModal = ({ selectedRefrigerator, setOpenModal }) => {
                 <label htmlFor="edit-refri-name-input" className='input-label'>Nombre de la heladera</label>
                 <input type="text" id='edit-refri-name-input' defaultValue={selectedRefrigerator.refri_name}  onChange={handleNameChange}/>
                 <label htmlFor="edit-refri-capacity-input" className='input-label'>Capacidad (baldes totales)</label>
-                <input type="number" id='edit-refri-capacity-input' defaultValue={selectedRefrigerator.total_capacity} min={6} max={32} step={2} onChange={handleAmountChange} /> 
+                <input type="number" id='edit-refri-capacity-input' defaultValue={selectedRefrigerator.total_capacity} min={6} max={32} step={2} onChange={handleAmountChange} disabled={!user} style={{cursor: !user && 'not-allowed'}} /> 
                 <small className='edit-modal-warn' hidden={hideWarning}>Advertencia: reducir la capacidad del refrigerador eliminará los baldes sobrantes</small>   
             </form>
             <div className="modal-btn-container">
